Separate Header and Rows into components
