+++ react/components/EXPERIMENTAL_Table/hooks/useTableState.ts
@@ -9,7 +9,10 @@ interface Input {
 }
 
 const useTableState = ({ columns, items }: Input): TableState => {
-  const isEmpty = useMemo(() => Object.keys(columns).length === 0, [columns])
+  const isEmpty = useMemo(
+    () => items.length === 0 || Object.keys(columns).length === 0,
+    [columns, items]
+  )
   const tableHeight = useMemo(
     () => calculateTableHeight(constants.ROW_HEIGHT, items.length),
     [items]
